Pulled chat into separate components
